Better fix for collation error in select query.
Instead of sending a SET NAMES query when creating the connection, do a
conversion of the column value returned from the JSON_TABLE call to the
utf8mb4 character set using its default collation. For some reason, this
is necessary or else that column ends up with a different collation.